refactor: Split operator_edge_cases.rs into focused modules
Refactored the 728-line operator_edge_cases.rs into focused modules:
- unary_operators.rs: Unary plus/minus tests
- binary_arithmetic.rs: Complex nested arithmetic
- comparison_operators.rs: Nested comparisons and logical ops
- null_propagation.rs: NULL handling across operators
- string_concatenation.rs: String concat edge cases
- operator_test_utils.rs: Shared test utilities

Benefits:
- Eliminated ~500 lines of duplicated SelectStmt construction
- Main file reduced from 728 lines to 21 lines
- Each module focuses on one operator category
- Shared utilities reduce boilerplate by 80%+
- Easier to add new operator edge case tests
- Better isolation for debugging specific operator issues

All 302 tests pass. No breaking changes to test execution or coverage.
